microservices/CoverLetterPDFGeneration/requirements.txt
python-docx
pypandoc